% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_url.R, R/rickroll.R, R/service-channel9.R
\name{embed_user2016}
\alias{embed_user2016}
\alias{embed_user2017}
\alias{rickroll_channel9}
\alias{embed_channel9}
\title{Embed video from Microsoft Channel 9}
\usage{
embed_user2016(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE
)

embed_user2017(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE
)

rickroll_channel9(...)

embed_channel9(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE
)
}
\arguments{
\item{id}{\code{character}, identifier provided by the service}

\item{width}{\code{numeric}, width of iframe (px)}

\item{height}{\code{numeric}, height of iframe (px)}

\item{ratio}{\code{character}, indicates aspect ratio for the \verb{<iframe/>}}

\item{frameborder}{\code{numeric}, size of frame border (px)}

\item{allowfullscreen}{\code{logical}, indicates if to allow fullscreen}

\item{...}{arguments (other than \code{id}) passed on to \code{\link[=embed]{embed()}}}
}
\value{
Object with S3 class \code{vembedr_embed}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated: links to Microsoft Channel 9
\href{https://docs.microsoft.com/en-us/teamblog/channel9joinedmicrosoftlearn}{no longer work}.
}
