% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_url.R
\name{embed_url}
\alias{embed_url}
\title{Embed video based on URL}
\usage{
embed_url(url)
}
\arguments{
\item{url}{\code{character}, URL of web-page for video}
}
\value{
Object with S3 class \code{vembedr_embed}.
}
\description{
You can use this function to embed video using only the URL and you do not
need any customization beyond the start-time. It works for all the
services supported by the \code{\link[=embed]{embed()}} family of functions.
}
\details{
This function calls \code{\link[=suggest_embed]{suggest_embed()}} then parses and evaluates the code.
}
\examples{
embed_url("https://youtu.be/1-vcErOPofQ?t=28s")
}
\seealso{
\code{\link[=suggest_embed]{suggest_embed()}}
}
