% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed_vimeo}
\alias{embed_vimeo}
\alias{embed_youtube}
\title{Creates an iframe for video-embedding}
\usage{
embed_vimeo(id, width = 500, height = 281, frameborder = 0,
  allow_full_screen = TRUE)

embed_youtube(id, width = 420, height = 315, frameborder = 0,
  allow_full_screen = TRUE)
}
\arguments{
\item{id}{character, url identifier}

\item{width}{numeric, width of iframe (px)}

\item{height}{numeric, height of iframe (px)}

\item{frameborder}{numeric, size of frame border (px)}

\item{allow_full_screen}{logical, indicates if to allow fullscreen}
}
\value{
html code describing iframe
}
\description{
Creates an iframe for video-embedding
}
\examples{
embed_vimeo("45196609")

embed_youtube("dQw4w9WgXcQ")

}

