% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_service}
\alias{get_service}
\title{Determine service based on URL}
\usage{
get_service(url)
}
\arguments{
\item{url}{\code{character}, URL of web-page for video}
}
\value{
\code{character} identifying the video service
}
\description{
Determine service based on URL
}
\examples{
get_service("https://youtu.be/1-vcErOPofQ?t=28s")
}
