% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iframe.R
\name{iframe}
\alias{iframe}
\alias{get_iframe}
\alias{get_iframe.default}
\alias{get_iframe.vembedr_embed}
\alias{set_iframe}
\alias{set_iframe.default}
\alias{set_iframe.vembedr_embed}
\title{Accessor methods for iframe}
\usage{
get_iframe(embed, ...)

\method{get_iframe}{default}(embed, ...)

\method{get_iframe}{vembedr_embed}(embed, ...)

set_iframe(embed, ...)

\method{set_iframe}{default}(embed, ...)

\method{set_iframe}{vembedr_embed}(embed, iframe, ...)
}
\arguments{
\item{embed}{\code{vembedr_embed} object, created using an \code{\link[=embed]{embed()}} function}

\item{...}{other args as needed}

\item{iframe}{\code{shiny.tag} with \code{name} \code{"iframe"}, object to be set}
}
\value{
\describe{
\item{get_iframe}{\code{shiny.tag} with \code{name} \code{"iframe"}, created using \link[htmltools:builder]{htmltools::tags}\verb{$iframe()}}
\item{set_iframe}{\code{embed} object}
}
}
\description{
Accessor methods for iframe
}
\keyword{internal}
