% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe functions}
\arguments{
\item{lhs, rhs}{An embed object and a fuction to apply to it}
}
\description{
Like dplyr, vembedr also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
use_start_time(rickroll_youtube(), "1m35s")
# you can write
rickroll_youtube() \%>\% use_start_time("1m35s")
}
