% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickroll.R
\name{rickroll_vimeo}
\alias{rickroll_vimeo}
\alias{rickroll_youtube}
\title{Embeds a popular video}
\usage{
rickroll_vimeo(...)

rickroll_youtube(...)
}
\arguments{
\item{...}{arguments passed on to \code{embed_foo}}
}
\value{
html code describing iframe
}
\description{
Embeds a popular video
}
\examples{
  rickroll_vimeo()
  rickroll_youtube()

}

