% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{secs}
\alias{secs}
\title{Get number of seconds given a string}
\usage{
secs(x)
}
\arguments{
\item{x}{character, describes a time duration, i.e. "3m15s"}
}
\value{
numeric, number of seconds
}
\description{
This is a helper function to get the number of seconds.
}
\details{
This could be useful for composing query parameters for YouTube embeds.
}
\examples{
  secs("45s")
  secs("3m15s")
  secs("1h1m5s")

}
\seealso{
\code{\link{embed_youtube}}, \code{\link{hms}}
}

