% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{suggest_embed}
\alias{suggest_embed}
\title{Suggest embedding-code based on a URL}
\usage{
suggest_embed(url, quiet = FALSE)
}
\arguments{
\item{url}{character, can be copied from browser location or from
the "share" output on a video's web page}

\item{quiet}{logical, indicates to supress printing of the suggested code}
}
\value{
character, silently returns the suggested code
}
\description{
This function is meant to work with URLs from any of the supported services.
}
\examples{
suggest_embed("https://youtu.be/1-vcErOPofQ?t=28s")
suggest_embed("https://www.youtube.com/watch?v=1-vcErOPofQ")
}

