% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{use_align}
\alias{use_align}
\title{Align horizontally}
\usage{
use_align(embed, align = c("left", "right", "center", "justified"))
}
\arguments{
\item{embed}{\code{vembedr_embed} object, created using an \code{\link[=embed]{embed()}} function}

\item{align}{\code{character}, indicates type of alignment}
}
\value{
object with class \code{vembedr_embed},
prints to an 'HTML' \code{<div/>} that contains an \code{<iframe/>}
}
\description{
Use this function to specify the horizontal alignment of the \code{<iframe/>}
within the enclosing \code{</div>}.
}
