% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{use_rounded}
\alias{use_rounded}
\title{Make rounded corners}
\usage{
use_rounded(embed, radius = NULL)
}
\arguments{
\item{embed}{\code{vembedr_embed} object, created using an \code{\link[=embed]{embed()}} function}

\item{radius}{\code{numeric} or \code{character}, css property for the border-radius
for the \code{<iframe/>}. Numeric values will be interpreted as number of pixels.}
}
\value{
object with class \code{vembedr_embed},
prints to an 'HTML' \code{<div/>} that contains an \code{<iframe/>}
}
\description{
You can use this function to make rounded corners for the enclosing \code{</div>}.
}
