% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vembedr-package.R
\docType{package}
\name{vembedr-package}
\alias{vembedr-package}
\title{vembedr: Package for embedding video}
\description{
The vembedr package lets you embed video into your HTML pages for
these services:
}
\details{
\itemize{
\item YouTube
\item Vimeo
\item Box
\item Microsoft Stream
}

It provides two categories of functions:
\itemize{
\item \strong{embed} functions, to specify a video to embed:
e.g. \code{\link[=embed_youtube]{embed_youtube()}}, \code{\link[=embed_url]{embed_url()}}
\item \strong{use} functions, to modify the embedding:
e.g. \code{\link[=use_start_time]{use_start_time()}}, \code{\link[=use_rounded]{use_rounded()}}
}

You can use the pipe (\verb{|>} or \verb{\%>\%}) to chain \emph{embed function-calls} with
\emph{use function-calls}.
}
