% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{vembedr-s3-classes}
\alias{vembedr-s3-classes}
\title{vembedr S3 Classes}
\description{
We use S3 classes to distinguish an embed object, and to denote which service it uses.
}
\details{
\strong{\code{vembedr_embed}}
\itemize{
\item base class for all services
\item HTML \verb{<div>}
\item contains the embed code
}

\strong{\code{vembedr_embed_youtube}}
\strong{\code{vembedr_embed_youtube_short}}
\strong{\code{vembedr_embed_vimeo}}
\strong{\code{vembedr_embed_channel9}}
\strong{\code{vembedr_embed_box}}
}
