% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vembedr.R
\docType{package}
\name{vembedr}
\alias{vembedr}
\alias{vembedr-package}
\title{vembedr: A package for embedding video}
\description{
The vembedr package provides two categories of functions:
embed and rickroll, to embed videos from the services: YouTube and Vimeo.
}
\details{
The important functions are \code{\link{embed_vimeo}}
and \code{\link{embed_youtube}}, with a useful helper function for
specifying the start time of a YouTube video: \code{\link{secs}}.
}

