% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\alias{embed_vimeo}
\alias{embed_youtube}
\alias{embed_user2016}
\alias{embed_user2017}
\alias{embed_channel9}
\alias{embed_box}
\title{Embed a video into an HTML document}
\usage{
embed_vimeo(id, width = NULL, height = 300, ratio = c("16by9",
  "4by3"), frameborder = 0, allowfullscreen = TRUE, query = NULL,
  fragment = NULL)

embed_youtube(id, width = NULL, height = 300, ratio = c("16by9",
  "4by3"), frameborder = 0, allowfullscreen = TRUE, query = NULL)

embed_user2016(id, width = NULL, height = 300, ratio = c("16by9",
  "4by3"), frameborder = 0, allowfullscreen = TRUE)

embed_user2017(id, width = NULL, height = 300, ratio = c("16by9",
  "4by3"), frameborder = 0, allowfullscreen = TRUE)

embed_channel9(id, width = NULL, height = 300, ratio = c("16by9",
  "4by3"), frameborder = 0, allowfullscreen = TRUE)

embed_box(id, custom_domain = getOption("vembedr.box_custom_domain"),
  width = NULL, height = 300, ratio = c("16by9", "4by3"),
  frameborder = 0, allowfullscreen = TRUE)
}
\arguments{
\item{id}{character, identifier provided by the service}

\item{width}{numeric, width of iframe (px)}

\item{height}{numeric, height of iframe (px)}

\item{ratio}{\code{character}, indicates aspect ratio for the \code{<iframe/>}}

\item{frameborder}{numeric, size of frame border (px)}

\item{allowfullscreen}{logical, indicates if to allow fullscreen}

\item{query}{list of items to include in url-query string}

\item{fragment}{character, string to include as url-fragment}

\item{custom_domain}{character, (used by Box) name of Box-instance
to use. It can be useful to use \code{getOption("vembedr.box_custom_domain")}
if you are using a corporate instance of Box. If \code{NULL}, it will use
the standard Box instance.}
}
\value{
object with class \code{vembedr_embed},
prints to an 'HTML' \code{<div/>} that contains an \code{<iframe/>}
}
\description{
These functions are used to embed video into your \strong{rmarkdown} html-documents,
or into your \strong{shiny} apps. There are functions to embed from
YouTube, Vimeo, and Microsoft Channel 9 (who host the UseR! 2016 videos).
}
\details{
These services allow you to customize a lot of things by specifying
an optional query string. The specification for the query string will differ
according to the service being used:

\describe{
\item{YouTube}{\url{https://developers.google.com/youtube/player_parameters}}
\item{Vimeo}{\url{https://developer.vimeo.com/player/embedding}}
\item{Channel 9}{\url{https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Forty-years-of-S}
(click the \emph{Embed} tab)}
\item{Box}{\url{https://developer.box.com/docs/box-embed#section-build-box-embed-programatically}}
}
}
\examples{
embed_youtube("dQw4w9WgXcQ")
embed_vimeo("45196609")
embed_user2016("Literate-Programming")
embed_user2017("Room-202-Lightning-Talks") \%>\% use_start_time("26m35s")
embed_box("m5do45hvzw32iv2aors3urf5pgkxxazx")

}
\seealso{
\code{\link[=use_start_time]{use_start_time()}}
}
\concept{embed}
