% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed_vimeo}
\alias{embed_vimeo}
\alias{embed_youtube}
\title{Creates an iframe for video-embedding}
\usage{
embed_vimeo(id, width = 500, height = 281)

embed_youtube(id, width = 500, height = 281)
}
\arguments{
\item{id}{character, vimeo_id}

\item{width}{numeric, width of iframe (pixels)}

\item{height}{numeric, height of iframe (pixels)}
}
\value{
html code describing iframe
}
\description{
Creates an iframe for video-embedding
}
\examples{
embed_vimeo(45196609)

embed_youtube("dQw4w9WgXcQ")

}

