% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_service}
\alias{get_service}
\title{Determine the service, given the URL}
\usage{
get_service(url)
}
\arguments{
\item{url}{character, can be copied from browser location or from
the "share" output on a video's web page}
}
\value{
\code{character} identfying the video service
}
\description{
Determine the service, given the URL
}
\examples{
get_service("https://youtu.be/1-vcErOPofQ?t=28s")
}
