% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hms}
\alias{hms}
\title{creates an hours-minutes-seconds string}
\usage{
hms(x)
}
\arguments{
\item{x, }{numeric (number of seconds), or character (i.e. "3m15s")}
}
\value{
character string (i.e. "0h3m15s")
}
\description{
creates an hours-minutes-seconds string
}
\examples{
  hms(30)
  hms("3m15s")

}
\seealso{
\code{\link{secs}}
}

