% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_video_url}
\alias{parse_video_url}
\title{Parse a URL to determine service and id}
\usage{
parse_video_url(url)
}
\arguments{
\item{url}{character, URL to parse}
}
\value{
list with members:
\describe{
\item{service}{character, describes which service is used}
\item{id}{character, identifier for the video at the service}
\item{start_time}{character, indicates start time}
}
}
\description{
This is an internal function, supporting \code{\link[=suggest_embed]{suggest_embed()}}
}
\examples{
parse_video_url("https://youtu.be/1-vcErOPofQ?t=28s")
}
\seealso{
suggest_embed
}
\keyword{internal}
