% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickroll.R
\name{rickroll}
\alias{rickroll}
\alias{rickroll_channel9}
\alias{rickroll_vimeo}
\alias{rickroll_youtube}
\title{Embeds a popular video}
\usage{
rickroll_vimeo(...)

rickroll_youtube(...)

rickroll_channel9(...)
}
\arguments{
\item{...}{arguments (other than \code{id}) passed on to \code{\link{embed}}}
}
\value{
html \code{<iframe>} element
}
\description{
If you want to experiment with the arguments to \code{\link{embed}},
such as \code{query}, but do not have a particular video in mind, this function
may be useful to you.
}
\examples{
  rickroll_vimeo()
  rickroll_youtube()
  rickroll_channel9() \%>\% use_start_time("5m03s")

}

