% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{suggest_embed}
\alias{suggest_embed}
\title{Suggest embedding-code based on a URL}
\usage{
suggest_embed(url)
}
\arguments{
\item{url}{character, can be copied from browser location or from
the "share" output on a video's web page}
}
\value{
called for side-effects,
  also silently returns the suggested code as character
}
\description{
Suggest embedding-code based on a URL
}
\examples{
suggest_embed("https://youtu.be/1-vcErOPofQ?t=28s")
suggest_embed("https://www.youtube.com/watch?v=1-vcErOPofQ")
}

