% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{use_bs_responsive}
\alias{use_bs_responsive}
\title{Make size responsive}
\usage{
use_bs_responsive(embed)
}
\arguments{
\item{embed}{\code{vembedr_embed} object, created using an \code{\link[=embed]{embed()}} function}
}
\value{
object with class \code{vembedr_embed},
prints to an 'HTML' \verb{<div/>} that contains an \verb{<iframe/>}
}
\description{
If your 'HTML\verb{page includes [Twitter Bootstrap 3](https://getbootstrap.com/docs/3.3/components/#responsive-embed), you can use this function to make the size of the}iframe\verb{responsive within the enclosing}div`.
}
