% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{use_bs_responsive}
\alias{use_bs_responsive}
\title{Make size responsive}
\usage{
use_bs_responsive(embed)
}
\arguments{
\item{embed}{\code{vembedr_embed} object, created using an \code{\link[=embed]{embed()}} function}
}
\value{
Object with S3 class \code{vembedr_embed}
}
\description{
If your HTML page includes
\href{https://getbootstrap.com/docs/3.3/components/#responsive-embed}{Twitter Bootstrap 3},
you can use this function to make the size of the \code{iframe} responsive
within the enclosing \code{div}.
}
