% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{vembedr-s3-classes}
\alias{vembedr-s3-classes}
\title{vembedr S3 Classes}
\description{
Knowledge of these classes is not needed for day-to-day use. Rather,
it is a bookkeeping device used to make it clearer to a developer
how to add a new service.
}
\details{
We use S3 classes to distinguish an embed object, and to denote which
service it uses. Objects of these classes are created by \code{\link[=embed_url]{embed_url()}}
and each service's embed function.

\strong{\code{vembedr_embed}}
\itemize{
\item base class for all services
\item HTML \verb{<div>}
\item contains the embed code
}

There is an additional class attached according to the service:
\itemize{
\item \strong{\code{vembedr_embed_youtube}}
\item \strong{\code{vembedr_embed_youtube_short}}
\item \strong{\code{vembedr_embed_vimeo}}
\item \strong{\code{vembedr_embed_channel9}}
\item \strong{\code{vembedr_embed_box}}
\item \strong{\code{vembedr_embed_msstream}}
}

To support parsing, there is an internal S3 class attached to the URL
being processed. It is named according to the service:
\itemize{
\item \strong{\code{vembedr_url_youtube}}
\item \strong{\code{vembedr_url_youtube_short}}
\item \strong{\code{vembedr_url_vimeo}}
\item \strong{\code{vembedr_url_channel9}}
\item \strong{\code{vembedr_url_box}}
\item \strong{\code{vembedr_url_msstream}}
}
}
