% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vembedr.R
\docType{package}
\name{vembedr}
\alias{vembedr}
\alias{vembedr-package}
\title{vembedr: A package for embedding video}
\description{
The vembedr package provides two categories of functions:
embed and rickroll, to embed videos from the services YouTube, Vimeo, and Microsoft Channel 9
(including UseR! 2016 and 2017 videos).
}
\details{
The important functions are \code{\link{embed_vimeo}}, \code{\link{embed_youtube}}, and \code{\link{embed_user2016}},
with a helper function for specifying the start times of videos: \code{\link{use_start_time}}.
}
